backend/django/drf/terceiro_curso_drf/projeto_inicial_drf_clientes-master/requirements.txt
asgiref==3.5.2
astroid==2.4.2
backports.zoneinfo==0.2.1;python_version<"3.9"
dj-database-url==1.0.0
Django==3.2.10
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.11.0
Faker==15.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.9.3
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.12.1
